(refactor) Split data streaming file into helper file and script
